rbuild: No longer fetches remote archives if they exist locally already
git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@6306 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Tools/Dev/rbuild/askapdev/rbuild/builders/builder.py
@@ -159,11 +159,13 @@ class Builder:
         if self.remote_archive is None:
             return
         uitem = urllib2.urlparse.urlsplit(self.remote_archive)
+        outfile = os.path.split(uitem.path)[-1]
+        if (os.path.isfile(outfile)):
+            return
         fullpath = self.remote_archive
         if not uitem.scheme:
             root = os.environ["RBUILD_REMOTE_ARCHIVE"]
             fullpath = os.path.sep.join((root, self.remote_archive))
-        outfile = os.path.split(uitem.path)[-1]
         remote = urllib2.urlopen(fullpath)
         utils.q_print("info: Fetching '{}'...".format(fullpath))
         with open(outfile, "wb") as of:
